forks now mutex

+++ time.c
@@ -6,23 +6,23 @@
 /*   By: ablanco- <ablanco-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/20 21:33:14 by ablanco-          #+#    #+#             */
-/*   Updated: 2023/12/20 21:41:55 by ablanco-         ###   ########.fr       */
+/*   Updated: 2024/01/02 20:26:15 by ablanco-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long get_time(void)
+int get_time(void)
 {
-	struct timeval start_time;
-	long ms;
+	struct timeval	start_time;
+	int				ms;
 
 	gettimeofday(&start_time, NULL);
 	ms = (start_time.tv_sec * 1000) + (start_time.tv_usec / 1000);
 	return(ms);
 }
 
-long dif_time(long start, long now)
+int dif_time(int start, int now)
 {
 	return (now - start);
 }
